chapter06/src/main/java/com/zgy/study/chapter06/subject05/demo01/DbConfig.java
package com.zgy.study.chapter06.subject05.demo01;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.context.annotation.*;
import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
import org.springframework.jdbc.datasource.SimpleDriverDataSource;

import javax.sql.DataSource;
import java.sql.Driver;

/**
 * @author ZGY <br>
 * @date 2019/11/29 15:06 <br>
 * @description DbConfig <br>
 */
@Configuration
@PropertySource("classpath:subject05/demo01/jdbc2.properties")
public class DbConfig {

    @Value("${jdbc.driverClassName}")
    private String driverClassName;

    @Value("${jdbc.url}")
    private String url;

    @Value("${jdbc.username}")
    private String username;

    @Value("${jdbc.password}")
    private String password;

    @Bean
    public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
        return new PropertySourcesPlaceholderConfigurer();
    }

    @Bean
    @Lazy
    public DataSource dataSource() {
        SimpleDriverDataSource dataSource = new SimpleDriverDataSource();
        try {
            Class<? extends Driver> driverClass = (Class<? extends Driver>) Class.forName(driverClassName);
            dataSource.setDriverClass(driverClass);
            dataSource.setUrl(url);
            dataSource.setUsername(username);
            dataSource.setPassword(password);

            return dataSource;
        } catch (ClassNotFoundException e) {
            return null;
        }
    }
}
